+++ plone/protect/monkey.py
@@ -1,4 +1,5 @@
 from urlparse import urlparse, urljoin
+from plone.protect.auto import safeWrite
 
 
 def RedirectTo__call__(self, controller_state):
@@ -28,10 +29,10 @@ def wl_lockmapping(self, killinvalids=0, create=0):
     has_write_locks = hasattr(self, '_dav_writelocks')
     locks = self._old_wl_lockmapping(killinvalids=killinvalids, create=create)
     try:
-        locks._v_safe_write = True  # hint to tell plone.protect to ignore this object
+        safeWrite(locks)
         if not has_write_locks and create:
             # first time writing to object, need to mark it safe
-            self._v_safe_write = True
+            safeWrite(self)
     except AttributeError:
         # not a persistent class, ignore
         pass